+++ priv/static/phoenix_live_view.cjs.js
@@ -109,11 +109,15 @@ var EntryUploader = class {
     this.offset = 0;
     this.chunkSize = chunkSize;
     this.chunkTimer = null;
+    this.errored = false;
     this.uploadChannel = liveSocket.channel(`lvu:${entry.ref}`, { token: entry.metadata() });
   }
   error(reason) {
+    if (this.errored) {
+      return;
+    }
+    this.errored = true;
     clearTimeout(this.chunkTimer);
-    this.uploadChannel.leave();
     this.entry.error(reason);
   }
   upload() {
@@ -145,7 +149,7 @@ var EntryUploader = class {
       if (!this.isDone()) {
         this.chunkTimer = setTimeout(() => this.readNextChunk(), this.liveSocket.getLatencySim() || 0);
       }
-    });
+    }).receive("error", ({ reason }) => this.error(reason));
   }
 };
 
@@ -312,7 +316,7 @@ var DOM = {
   wantsNewTab(e) {
     let wantsNewTab = e.ctrlKey || e.shiftKey || e.metaKey || e.button && e.button === 1;
     let isDownload = e.target instanceof HTMLAnchorElement && e.target.hasAttribute("download");
-    let isTargetBlank = e.target.getAttribute("target") === "_blank";
+    let isTargetBlank = e.target.hasAttribute("target") && e.target.getAttribute("target").toLowerCase() === "_blank";
     return wantsNewTab || isTargetBlank || isDownload;
   },
   isUnloadableFormSubmit(e) {
@@ -3562,7 +3566,8 @@ var View = class {
     let linkRef = this.liveSocket.setPendingLink(href);
     let refGen = targetEl ? () => this.putRef([targetEl], "click") : null;
     let fallback = () => this.liveSocket.redirect(window.location.href);
-    let push = this.pushWithReply(refGen, "live_patch", { url: href }, (resp) => {
+    let url = href.startsWith("/") ? `${location.protocol}//${location.host}${href}` : href;
+    let push = this.pushWithReply(refGen, "live_patch", { url }, (resp) => {
       this.liveSocket.requestDOMUpdate(() => {
         if (resp.link_redirect) {
           this.liveSocket.replaceMain(href, null, callback, linkRef);
@@ -4214,6 +4219,7 @@ var LiveSocket = class {
       }
       let { type, id, root, scroll } = event.state || {};
       let href = window.location.href;
+      dom_default.dispatchEvent(window, "phx:navigate", { detail: { href, patch: type === "patch", pop: true } });
       this.requestDOMUpdate(() => {
         if (this.main.isConnected() && (type === "patch" && id === this.main.id)) {
           this.main.pushLinkPatch(href, null, () => {
@@ -4291,6 +4297,7 @@ var LiveSocket = class {
       return;
     }
     browser_default.pushState(linkState, { type: "patch", id: this.main.id }, href);
+    dom_default.dispatchEvent(window, "phx:navigate", { detail: { patch: true, href, pop: false } });
     this.registerNewLocation(window.location);
   }
   historyRedirect(href, linkState, flash) {
@@ -4305,6 +4312,7 @@ var LiveSocket = class {
     this.withPageLoading({ to: href, kind: "redirect" }, (done) => {
       this.replaceMain(href, flash, () => {
         browser_default.pushState(linkState, { type: "redirect", id: this.main.id, scroll }, href);
+        dom_default.dispatchEvent(window, "phx:navigate", { detail: { href, patch: false, pop: false } });
         this.registerNewLocation(window.location);
         done();
       });